refactor(core): Split validActions and helpers into focused modules for LLM optimization
validActions/units.ts (539 lines) → units/:
- recruitment.ts: Reputation cost, site access, recruitable units
- activation.ts: Ability validation, phase checks, combat restrictions

validActions/index.ts (537 lines) → extracted helpers:
- pending.ts: Glade wound and deep mine choice options
- tactics.ts: Tactic selection and effect options

helpers/rewardHelpers.ts (469 lines) → rewards/:
- types.ts: RewardResult type and die face constants
- queueing.ts: Queue rewards for later selection
- handlers.ts: Individual reward type handlers

Reduces orchestrator files to focused modules for better LLM context efficiency.
