+++ packages/core/src/engine/validActions/units/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Unit valid actions - recruitment and activation options
+ *
+ * Split into focused modules:
+ * - recruitment: Site-based unit recruitment options
+ * - activation: Unit ability activation options
+ */
+
+export {
+  getReputationCostModifier,
+  siteTypeToRecruitSite,
+  isSiteAccessibleForRecruitment,
+  getUsedCommandTokens,
+  getUnitOptions,
+} from "./recruitment.js";
+
+export {
+  getActivatableUnits,
+  getUnitOptionsForCombat,
+  getFullUnitOptions,
+} from "./activation.js";
